+++ hftbacktest/src/connector/mod.rs
@@ -17,8 +17,8 @@ pub trait Connector {
         &mut self,
         asset_no: usize,
         symbol: String,
-        tick_size: f32,
-        lot_size: f32,
+        tick_size: f64,
+        lot_size: f64,
     ) -> Result<(), anyhow::Error>;
 
     /// Runs the connector, establishing the connection and preparing to exchange information such
